requirements.txt
pypdf>=3.0.0
python-dotenv>=1.0.0
openai>=1.0.0
pandas>=2.0.0
tiktoken>=0.5.0

